Add files via upload
Custom child theme to create a search box to pull data from MongoDB via API. Data will be shown in a dropdown manner. User can click on an individual item to see the details or the user can click on More link to see the listing page. Free user can use this search box for 30-times after that he/she needs to register on the site to use the search box.